Run tests as a non-prived user and avoid kcov limitation (BREAKING)
kcov depends on the PS4 variable, which due to a bash security fix, no
longer works when bash is run as root:

https://github.com/SimonKagstrom/kcov/issues/234

This results in no coverage analysis when run on a recent Docker base
image that contains the bash security fix.

Running as a non-prived user in Docker works around the limitation.

This is a breaking changes on projects downstream, which need to adapt
their Dockerfiles to run tests as the non-prived "bashtestdummy" user,
yet install dependencies as root.  `example/tests/Dockerfile` has
been updated to match.